Remove TransactionBuilder from tests (besides transaction_builder.spec.ts)
